+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 
 ## Unreleased
 - added fail on missing service flag
+- ignore checks during initialization flag
 
 ## [0.0.5] - 2015-11-26
 ### Changed